fix: calculateForInvoiceItem resolves the article tax group (AID-138) (#39)
TaxCalculationService::calculateForInvoiceItem was a stub — the tax
lookup was commented out as TODO, so it always returned zero tax and no
taxes_applied. InvoiceService::createInvoice therefore produced invoices
without VAT, and downstream incorrect exempt Verifactu breakdowns.

- Resolve the tax group: explicit tax_group_id wins, otherwise the
  article's tax_group_id; delegate to the configured strategy. Document
  the base-100 units contract (qty=100 means 1.0 unit).
- InvoiceService::createInvoiceItem persists the immutable taxes_applied
  snapshot and article_id, and forwards a per-item tax_group_id override.
- VatCalculationStrategy emits integer tax amounts (base-100 invariant);
  it leaked floats from round() into the snapshot. Existing strategy
  tests asserted the accidental floats and were aligned to ints.

+++ src/Services/TaxCalculation/VatCalculationStrategy.php
@@ -16,7 +16,10 @@ class VatCalculationStrategy implements TaxCalculationStrategy
 
         foreach ($taxGroup->taxRates as $taxRate) {
             // Lógica de cálculo de IVA simple por ahora. Se puede expandir.
-            $taxAmount = round($baseAmount * ($taxRate->rate / 10000));
+            // Int cast: monetary values are base-100 integers across the
+            // ecosystem (lara100 invariant), including the immutable
+            // taxes_applied snapshot persisted on invoice items.
+            $taxAmount = (int) round($baseAmount * ($taxRate->rate / 10000));
 
             $totalTaxAmount += $taxAmount;
             $taxesApplied[] = [
